+++ master.tf
@@ -1,5 +1,5 @@
 resource "aws_s3_bucket" "bucket" {
-  count  = var.is_guardduty_master && (var.has_ipset || var.has_threatintelset) ? 1 : 0
+  count  = var.is_guardduty_master && (length(var.ipset_iplist) > 0 || length(var.threatintelset_iplist) > 0) ? 1 : 0
   bucket = var.bucket_name
   versioning {
     enabled = true
@@ -9,7 +9,7 @@ resource "aws_s3_bucket" "bucket" {
 }
 
 resource "aws_s3_bucket_object" "ipset" {
-  count = var.is_guardduty_master && var.has_ipset ? 1 : 0
+  count = var.is_guardduty_master && length(var.ipset_iplist) > 0 ? 1 : 0
   acl   = "public-read"
   content = templatefile("${path.module}/templates/ipset.txt.tpl",
   { ipset_iplist = var.ipset_iplist })
@@ -18,8 +18,8 @@ resource "aws_s3_bucket_object" "ipset" {
 }
 
 resource "aws_guardduty_ipset" "ipset" {
-  count       = var.is_guardduty_master && var.has_ipset ? 1 : 0
-  activate    = var.ipset_activate
+  count       = var.is_guardduty_master && length(var.ipset_iplist) > 0 ? 1 : 0
+  activate    = true
   detector_id = aws_guardduty_detector.detector.id
   format      = var.ipset_format
   location    = "https://s3.amazonaws.com/${aws_s3_bucket.bucket[0].id}/${local.ipset_key}"
@@ -27,7 +27,7 @@ resource "aws_guardduty_ipset" "ipset" {
 }
 
 resource "aws_s3_bucket_object" "threatintelset" {
-  count = var.is_guardduty_master && var.has_threatintelset ? 1 : 0
+  count = var.is_guardduty_master && length(var.threatintelset_iplist) > 0 ? 1 : 0
   acl   = "public-read"
   content = templatefile("${path.module}/templates/threatintelset.txt.tpl",
   { threatintelset_iplist = var.threatintelset_iplist })
@@ -36,8 +36,8 @@ resource "aws_s3_bucket_object" "threatintelset" {
 }
 
 resource "aws_guardduty_threatintelset" "threatintelset" {
-  count       = var.is_guardduty_master && var.has_threatintelset ? 1 : 0
-  activate    = var.threatintelset_activate
+  count       = var.is_guardduty_master && length(var.threatintelset_iplist) > 0 ? 1 : 0
+  activate    = true
   detector_id = aws_guardduty_detector.detector.id
   format      = var.threatintelset_format
   location    = "https://s3.amazonaws.com/${aws_s3_bucket.bucket[0].id}/${local.threatintelset_key}"