+++ CHANGELOG.md
@@ -5,7 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [0.3.0] 2020-03-30
-- Aligned with 3 Musketeers pattern
+### Changed
+- Changed structure of resource, dependencies and tests to align with standards
 
 ## [0.2.0] 2020-01-16
 - Give meaningful name for bucket used by GuardDuty @jason.lau